fix: prob6 type convert to int

+++ Prob6_type_conversion.go
@@ -5,5 +5,5 @@ import "fmt"
 func main () {
 	accountAge:= 2.5
 
-	fmt.Println("Your account has existed for ",accountAge,"years.")
+	fmt.Println("Your account has existed for ", uint(accountAge),"years.")
 }
